+++ GAM200/src/LevelEditor.cpp
@@ -1071,7 +1071,7 @@ void LevelEditor::ListOfObjects() {
 	if (ImGui::Button("Create new layer")) {
 		std::string layerName = std::string("Layer " + static_cast<char>(sceneManager->layers.size()));
 		//buffer
-		sprintf_s(buffer, "Layer %d", sceneManager->layers.size());
+		sprintf_s(buffer, "Layer %d", static_cast<int>(sceneManager->layers.size()));
 		objectFactory->CreateLayer(std::string(buffer), true);
 	}
 	ImGui::BeginChild("ObjectListScroll", ImGui::GetContentRegionAvail());
